fix problem in partitioning precipitation between canopy and soil
`phi_t` was used in Artemis as a "throughfall coefficient", but the equation originates from [Hough and Jones (1997)](https://doi.org/10.5194/hess-1-227-1997) where it is calculating the "proportion of the rainfall intercepted" (i.e. the complementary value for the "throughfall coefficient"). This is fixed by simply swapping round `p_can` and `p_soil` to take this into account.

In addition, the fix I had added for the case where LAI=0 using `np.ma.where` is removed as it was wrong and creating an unjustified discontinuity in the function at 0.