add special case for LAI=0 for evaporation from canopy store
When LAI=0, the calculation for evaporation from the canopy store was resulting in a ZeroDivisionError.

When LAI=0, the canopy store becomes effectively of capacity zero (i.e. `C_t = 0`), meaning that there cannot be evaporation from the store since it cannot contain any water. So in that case, `e_can` is set to zero. In the algorithm, the water previously present in the canopy store will become throughfall (`q_t`), which can be assumed to be physically realistic as the water on the canopy leaves will fall to the ground with the leaves.