adjust z origin for wind to avoid log of negative value in aerodynamic resistance calculation
When zero plane displacement (calculated as a fraction on vegetation height) gets larger than the measurement height for wind speed, the calculation for aerodynamic resistance results in computing the log of negative values.

To avoid this problem, the same approach as in JULES is taken, that is to say that the origin of the vertical dimension is taken as where the wind profile gets to zero (that is to say at a height equals to the zero-plane displacement + the roughness length). Then, it is assuming that the wind speed measurement (e.g. 10m) is from this origin rather than the ground. This is not physically realistic, but this gets around the problem when vegetation height is such that zero-plane displacement is above the measurement height for wind speed (or humidity).

The scaling of the wind speed to 2m is also removed as the wind speed value must be consistent with the value used for the measurement heights which is not 2m, but varies depending on the value of `z`. The value of `z` is expected to be changed by the user to match the wind speed measurements used to drive the component.